create the post type registrar
